enable spatial parallelism

+++ fme/core/generics/trainer.py
@@ -386,8 +386,8 @@ class Trainer:
             wandb = WandB.get_instance()
             wandb.log(all_logs, step=self.num_batches_seen)
 
-            if dist.is_root():
-                if self.config.save_checkpoint:
+            if self.config.save_checkpoint:
+                if dist.is_root():
                     logging.info(f"Saving checkpoints for epoch {self._epochs_trained}")
                     self.save_all_checkpoints(valid_loss, inference_error)
 
@@ -417,7 +417,6 @@ class Trainer:
         )
         self.train_data.set_epoch(self._epochs_trained + 1)
         wandb = WandB.get_instance()
-        dist = Distributed.get_instance()
         names_to_log = ("batch_loss", "training_samples_per_second_on_rank_0", "lr")
         aggregator = self._aggregator_builder.get_train_aggregator()
         n_samples_seen_since_logging = 0
@@ -471,13 +470,12 @@ class Trainer:
                 logging.info(f"Step {self.num_batches_seen}: {metrics_to_log}")
                 n_samples_seen_since_logging = 0
             if (
-                dist.is_root()
-                and self.config.checkpoint_every_n_batches > 0
+                self.config.checkpoint_every_n_batches > 0
                 and self.num_batches_seen % self.config.checkpoint_every_n_batches == 0
             ):
                 self._save_restart_checkpoints()
                 self._last_saved_num_batches_seen = self.num_batches_seen
-        if dist.is_root() and self.num_batches_seen > self._last_saved_num_batches_seen:
+        if self.num_batches_seen > self._last_saved_num_batches_seen:
             self._save_restart_checkpoints()  # before incrementing epoch so we will validate after resuming  # noqa: E501
         # we will save restart checkpoints again after validation/inference
         # are recorded to wandb
@@ -582,8 +580,7 @@ class Trainer:
         ema_checkpoint_path: str | None = None,
         include_optimization: bool = False,
     ):
-        if not Distributed.get_instance().is_root():
-            raise RuntimeError("Only the root process should save checkpoints")
+        dist = Distributed.get_instance()
         # save to a temporary file in case we get pre-empted during save
         temporary_location = os.path.join(
             os.path.dirname(checkpoint_path), f".{uuid.uuid4()}.tmp"
@@ -617,14 +614,15 @@ class Trainer:
                     if "optimization" in ema_data:
                         ema_data.pop("optimization")
                     torch.save(ema_data, ema_temporary_location)
-            torch.save(data, temporary_location)
-            if ema_temporary_location is not None and ema_checkpoint_path is not None:
-                os.replace(ema_temporary_location, ema_checkpoint_path)
-            os.replace(temporary_location, checkpoint_path)
+            if dist.is_root():
+                torch.save(data, temporary_location)
+                if ema_temporary_location is not None and ema_checkpoint_path is not None:
+                    os.replace(ema_temporary_location, ema_checkpoint_path)
+                os.replace(temporary_location, checkpoint_path)
         finally:
-            if os.path.exists(temporary_location):
+            if dist.is_root() and os.path.exists(temporary_location):
                 os.remove(temporary_location)
-            if ema_temporary_location is not None and os.path.exists(
+            if dist.is_root() and ema_temporary_location is not None and os.path.exists(
                 ema_temporary_location
             ):
                 os.remove(ema_temporary_location)
